+++ ChangeLog.md
@@ -3,6 +3,7 @@ Data sequences change log
 
 ## ?.?.? / ????-??-??
 
+* Implemented util.data.Optional::toString() - @thekid
 * Implemented util.data.Sequence::toString() - @thekid
 
 ## 2.2.1 / 2015-05-31